fix: docker fixes, database initialization moved to libs/database, now every service has its own database
