+++ .env.example
@@ -3,14 +3,12 @@ GRPC_PORT=9001
 HTTP_PORT=8000
 GRPC_ADDR=scraper:9001
 
-
 # tg-bot
 TELEGRAM_TOKEN=token_from_@botfather
 
 # postgres
-POSTGRES_USER=postgres
 POSTGRES_PASSWORD=password
-POSTGRES_HOST=postgres
-POSTGRES_PORT=5432
+POSTGRES_USER=postgres
 POSTGRES_DB=postgres
 POSTGRES_SSL=disable
+POSTGRES_PORT=5432
