+++ docker-compose.yml
@@ -16,12 +16,17 @@ services:
         env_file:
             - .env
         depends_on:
-            - scraper
-            - nats
+            # scraper:
+            # condition: service_healthy #TODO healthcheck for scraper
+            nats:
+                condition: service_started
+            tg-bot-postgres:
+                condition: service_healthy
         restart: unless-stopped
         environment:
             - IS_DOCKER=true
             - GRPC_ADDR=scraper:${GRPC_PORT}
+            - POSTGRES_HOST=tg-bot-postgres
 
     scraper:
         build:
@@ -32,8 +37,10 @@ services:
         env_file:
             - .env
         depends_on:
-            - nats
-            - postgres
+            nats:
+                condition: service_started
+            scraper-postgres:
+                condition: service_healthy
         restart: unless-stopped
         expose:
             - '${GRPC_PORT:-9001}'
@@ -41,17 +48,34 @@ services:
             - '${HTTP_PORT:-8000}:${HTTP_PORT:-8000}'
         environment:
             - IS_DOCKER=true
+            - POSTGRES_HOST=scraper-postgres
 
-    postgres:
+    scraper-postgres:
         image: postgres
         container_name: scraper-postgres
         restart: unless-stopped
         volumes:
-            - ./migrations/:/docker-entrypoint-initdb.d/
+            - ./migrations/scraper:/docker-entrypoint-initdb.d/
         env_file:
             - .env
         ports:
             - '5433:5432'
+        healthcheck:
+            test: ['CMD-SHELL', 'pg_isready -U ${POSTGRES_USER}']
+            interval: 3s
+            timeout: 3s
+            retries: 5
+
+    tg-bot-postgres:
+        image: postgres
+        container_name: tg-bot-postgres
+        restart: unless-stopped
+        volumes:
+            - ./migrations/tg-bot:/docker-entrypoint-initdb.d/
+        env_file:
+            - .env
+        ports:
+            - '5434:5432'
         healthcheck:
             test: ['CMD-SHELL', 'pg_isready -U ${POSTGRES_USER}']
             interval: 3s
@@ -63,3 +87,5 @@ services:
         restart: unless-stopped
         ports:
             - '8080:8080'
+        environment:
+            - ADMINER_DEFAULT_SERVER=scraper-postgres,tg-bot-postgres
