--- services/tg-bot/internal/database/database.go
@@ -1,42 +1,0 @@
-package database
-
-import (
-	"context"
-	"fmt"
-	"os"
-
-	"github.com/jackc/pgx/v5/pgxpool"
-)
-
-type Database struct {
-	Pool *pgxpool.Pool
-}
-
-func New() (*Database, error) {
-	url := fmt.Sprintf(
-		"postgres://%s:%s@%s:%s/%s?sslmode=%s",
-		os.Getenv("POSTGRES_USER"),
-		os.Getenv("POSTGRES_PASSWORD"),
-		os.Getenv("POSTGRES_HOST"),
-		os.Getenv("POSTGRES_PORT"),
-		os.Getenv("POSTGRES_DB"),
-		os.Getenv("POSTGRES_SSL"),
-	)
-
-	pool, err := pgxpool.New(context.Background(), url)
-	if err != nil {
-		return nil, fmt.Errorf("db: connect to database: url: %s, err: %w", url, err)
-	}
-
-	return &Database{
-		Pool: pool,
-	}, nil
-}
-
-func (d *Database) Close() {
-	d.Pool.Close()
-}
-
-func (d *Database) Ping(ctx context.Context) error {
-	return d.Pool.Ping(ctx)
-}
